finished dz3 fine
