--- simple-chat/.gitignore
@@ -1,2 +1,0 @@
-simple-chat/node_modules
-simple-chat/build
