+++ simple-chat/src/chat_list.html
@@ -4,11 +4,12 @@
     <head lang="en">
         <meta charset="UTF-8">
         <title>Message list</title>
-        <link href="./styles/chat_list.css" rel="stylesheet">
+        <link href="index.css" rel="stylesheet"  type="text/css"/>
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
         <meta name="viewport" content="width=device-width, user-scalable=no">
     </head>
     <body> 
+        
             <div class = "header">
                 <button class="material-icons" >
                     <i class="material-icons">menu</i>
@@ -24,50 +25,43 @@
                 </div>
             </div>
             <div class="chat_field">
-                <div class="chat" onclick="window.location.href='./index.html';">
-                    <div class = "chat-im"> <img src = "./photos/Jennifer.jpg" /> </div >
+                <div class="chat" onclick="window.location.href='index.html';">
+                    <div class = "chat-im"> <img src = "https://sun6-23.userapi.com/s/v1/if1/2NuR7uGgIg0_KTiJTjGU9xN4Az8bTGXwhHW0uZ1Ls6d8N5ICSVI4ezbavb3zRR6DJQyL5A.jpg?size=667x667&quality=96&crop=165,0,667,667&ava=1" /> </div >
                     <div class = "mes-content">
                         <div class="message-name">Дженнифер</div>
                         <span class="message-text">Привет, Джен, всё нормально, просто на работе завал.</span>
                     </div>
-                    <div class="message-meta">
+                    <div class="message-meta1">
                             <span class="message-status">
                                 <i class="material-icons">done_all</i>
                             <span class="message-time">10:53</span>
                     </div>
                 </div>
-                <div class="chat" onclick="window.location.href='./yen.html';">
-                    <div class = "chat-im"> <img src = "./photos/Yennifer.jpg" /> </div >
+                <div class="chat" onclick="window.location.href='yen.html';">
+                    <div class = "chat-im"> <img src = "https://i.pinimg.com/736x/a9/d3/d0/a9d3d05e702fd49ab554e034288e8fd6.jpg" /> </div >
                     <div class = "mes-content">
-                        <div class="message-name">Йеннифер</div>
+                        <div class="message-name1">Йеннифер</div>
                         <span class="message-text">Привет, тебе нравится моя новая аватарка?.</span>
                     </div>
-                    <div class="message-meta">
+                    <div class="message-meta1">
                             <span class="message-time">8:34</span>
                             
                         
                     </div>
                 </div>
-                <div class="chat" onclick="window.location.href='./mike.html';">
-                    <div class = "chat-im"> <img src = "./photos/Mike.jpg" /> </div >
+                <div class="chat" onclick="window.location.href='mike.html';">
+                    <div class = "chat-im"> <img src = "https://st2.depositphotos.com/1773655/5632/i/450/depositphotos_56329841-stock-photo-young-man-pointing-up-showing.jpg" /> </div >
                     <div class = "mes-content">
                         <div class="message-name">Майк</div>
                         <span class="message-text">Том, у нас горят сроки, ускорься, пожалуйста.</span>
                     </div>
-                    <div class="message-meta">
+                    <div class="message-meta1">
                             <span class="message-time">8:24</span>
-                            
-                                
-                        
                     </div>
                 </div>
                 <button class="add_button" >
                     <i class="material-icons">add</i>
                 </button>
             </div>
-        
-        <script src="./scripts/chat_list.js"></script>
-                
-            
     </body>
 </html>
