+++ simple-chat/src/index.html
@@ -4,17 +4,17 @@
     <head lang="en">
         <meta charset="UTF-8">
         <title>Send a message</title>
-        <link href="./styles/index.css" rel="stylesheet">
+        <link href="index.css" rel="stylesheet"  type="text/css"/>
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
         <meta name="viewport" content="width=device-width, user-scalable=no">
     </head>
     <body>
 
             <div class = "head">
-                <button class="material-icons" onclick="window.location.href='/chat_list.html';">
+                <button class="material-icons" onclick="window.location.href='chat_list.html';">
                     <i class="material-icons">arrow_back</i>
                 </button>
-                <div class="img-head"><img src="./photos/Jennifer.jpg"></div>
+                <div class="img-head"><img src="https://sun6-23.userapi.com/s/v1/if1/2NuR7uGgIg0_KTiJTjGU9xN4Az8bTGXwhHW0uZ1Ls6d8N5ICSVI4ezbavb3zRR6DJQyL5A.jpg?size=667x667&quality=96&crop=165,0,667,667&ava=1"></div>
                 
                 <div class="status">
                     <h2>Дженнифер</h2>
@@ -34,14 +34,14 @@
 
 
                 <div class="message message-left">
-                    <div class = "btalk-im"> <img src = "./photos/Jennifer.jpg"  /> </div >
+                    <div class = "btalk-im"> <img src = "https://sun6-23.userapi.com/s/v1/if1/2NuR7uGgIg0_KTiJTjGU9xN4Az8bTGXwhHW0uZ1Ls6d8N5ICSVI4ezbavb3zRR6DJQyL5A.jpg?size=667x667&quality=96&crop=165,0,667,667&ava=1"  /> </div >
                     <span class="message-text">Том, почему тебя вчера не было, что-то случилось?</span>
                     <div class="message-meta">
                         <span class="message-time">10:53</span>
                     </div>
                 </div>
                 <div class="message message-right">
-                    <div class = "atalk-im"> <img src = "./photos/Tom.jpg" /> </div >
+                    <div class = "atalk-im"> <img src = "https://i.pinimg.com/originals/7a/26/8f/7a268f7ce7e63295d5c26922335adda9.jpg" /> </div >
                     <span class="message-text">Привет, Джен, всё нормально, просто на работе завал.</span>
                     <div class="message-meta">
                         <span class="message-time">10:53</span>
@@ -65,6 +65,6 @@
 
             </div>
        
-        <script src="./scripts/index.js"></script>
+        
     </body>
 </html>
