+++ simple-chat/src/mike.html
@@ -178,7 +178,7 @@
                 <button class="material-icons" onclick="window.location.href='/chat_list.html';">
                     <i class="material-icons">arrow_back</i>
                 </button>
-                <div class="img-head"><img src="/src/photos/Mike.jpg"></div>
+                <div class="img-head"><img src="https://st2.depositphotos.com/1773655/5632/i/450/depositphotos_56329841-stock-photo-young-man-pointing-up-showing.jpg"></div>
                 
                 <div class="status">
                     <h2>Майк</h2>
@@ -198,8 +198,8 @@
 
 
                 <div class="message message-left">
-                    <div class = "btalk-im"> <img src = "/src/photos/Mike.jpg"  /> </div >
-                    <span class="message-text">Привет, тебе нравится моя новая аватарка?</span>
+                    <div class = "btalk-im"> <img src = "https://st2.depositphotos.com/1773655/5632/i/450/depositphotos_56329841-stock-photo-young-man-pointing-up-showing.jpg"  /> </div >
+                    <span class="message-text">Том, у нас горят сроки, ускорься, пожалуйста.</span>
                     <div class="message-meta">
                         <span class="message-time">8:34</span>
                     </div>
@@ -293,7 +293,7 @@ function createDiv(message){
         new_message_img_div.className = "atalk-im";
         
         let new_message_img = document.createElement('img')
-        new_message_img.src = "/src/photos/Tom.jpg";
+        new_message_img.src = "https://i.pinimg.com/originals/7a/26/8f/7a268f7ce7e63295d5c26922335adda9.jpg";
         
         let new_message = document.createElement('div');
         new_message.className = "message message-right";
