--- simple-chat/src/styles/chat_list.css
@@ -1,140 +1,0 @@
-button{
-    border: 0;
-}
-
-.material-icons {
-    font-family: 'Material Icons';
-    font-weight: normal;
-    font-style: normal;
-    font-size: 36px;  /* Preferred icon size */
-    display: inline-block;
-    
-    text-transform: none;
-    letter-spacing: normal;
-    word-wrap: normal;
-    white-space: nowrap;
-    direction: ltr;
-    background-color: inherit;
-    width: 50px;
-    border: 0;
-    flex-basis: 60px;
-    
-}
-
-.material-icons1 {
-    font-family: 'Material Icons';
-    font-weight: normal;
-    font-style: normal;
-    font-size: 36px;  /* Preferred icon size */
-    display: inline-block;
-    
-    text-transform: none;
-    letter-spacing: normal;
-    word-wrap: normal;
-    white-space: nowrap;
-    direction: ltr;
-    background-color: inherit;
-    position: fixed; /* Фиксированное положение */
-    right: 10px; /* Расстояние от правого края окна браузера */
-     /* Расстояние сверху */
-    
-}
-
-.header{
-    width: 100%;
-    height: 12vh;
-    color: white;
-    display: flex;
-    background-color: blueviolet;
-    position: sticky;
-    top: 0;
-    align-items:center;
-}
-.status{
-    margin-left: 2%;
-    
-}
-
-body{
-    display: inline-block;
-    margin: 0;
-    width: 100%;
-    height: 100%;
-}
-
-.chat_field{
-    display: inline-block;
-    width: 100%;
-    height: 82vh;
-    flex-grow: 1;
-    background-color: white;
-    overflow-y: scroll;
-    align-items:center;
-}
-.add_button{
-    color: inherit;
-    position: absolute;
-    bottom: 5vh;
-    right: 5vh;
-    background-color: #ffca00ba;
-    height: 8vh;
-    width: 8vh;
-    display: flex;
-    align-content: center;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    box-shadow: 0px 5px 7px 1px rgb(0 0 0 / 20%);
-    transition: 1s ease-out;
-}
-.atalk-im{
-    border-radius: 50%;
-    width: 20px;
-    margin-left: 5%;
-    margin-top: 12px;
-    float: left;
-}
-.mes-content{
-    display: inline-block;
-    width: 160px;
-    margin-left: 3%;
-    align-items:center;
-    font-weight: normal;
-    margin-right: 20%;
-}
-
-.chat{
-    display: flex;
-    align-items:center;
-    
-}
-
-.message-meta{
-    display: inline-block;
-    width: 30px;
-    align-items:center;
-    position: fixed; /* Фиксированное положение */
-    right: 5vh; /* Расстояние от правого края окна браузера */
-    margin-left: 20px;
-    
-}
-
-
-.message-name{
-    font: 20px sans-serif;
-    margin-right: 120px;
-}
-.message-text{
-    font: 15px sans-serif;
-}
-.chat-im{
-    border-radius: 50%;
-    width: 100px;
-    margin-left: 5%;
-    margin-top: 12px;
-    float: left;
-}
-img{
-    max-width: 100%;
-    border-radius: 50%;
-}
