--- simple-chat/src/styles/index.css
@@ -1,152 +1,0 @@
-.material-icons {
-    font-family: 'Material Icons';
-    font-weight: normal;
-    font-style: normal;
-    font-size: 36px;  /* Preferred icon size */
-    display: inline-block;
-    
-    text-transform: none;
-    letter-spacing: normal;
-    word-wrap: normal;
-    white-space: nowrap;
-    direction: ltr;
-    background-color: inherit;
-    width: 50px;
-    border: 0;
-}
-
-img{
-    max-width: 100%;
-    border-radius: 50%;
-}
-.head{
-    width: 100%;
-    height: 12%;
-    color: white;
-    display: flex;
-    background-color: blueviolet;
-    position: sticky;
-    top: 0;
-    align-items:center;
-    
-    
-
-}
-
-.img-head{
-    border-radius: 50%;
-    width: 75px;
-    margin-left: 5%;
-    margin-top: 1%;
-    float: left;
-
-}
-.message {
-    display: flex;
-    width: fit-content;
-    border-radius: 5px;
-    border: 1px solid #bcb7be;
-    flex-direction: column;
-    row-gap: 20px;
-}
-.message-left{
-    padding: 12px 12px 0 12px;
-    margin: 7px 7px auto 7px;
-    border-bottom-left-radius: 0;
-    background-color: plum;
-}
-.message-right{
-    padding: 12px 12px 0 12px;
-    margin: 7px 7px 7px auto;
-    border-bottom-right-radius: 0;
-    background-color: wheat;
-    right: 0;
-}
-
-.atalk-im{
-    border-radius: 50%;
-    width: 20px;
-    margin-left: 5%;
-    margin-top: 12px;
-    float: left;
-}
-.btalk-im{
-    border-radius: 50%;
-    width: 20px;
-    margin-left: 5%;
-    margin-top: 12px;
-    float: left;
-}
-
-
-body{
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    height: 100%;
-}
-
-.status{
-    width: 120px;
-    margin-left: 3%;
-    font: 15px/27px sans-serif;
-    font-weight: normal;
-    margin-right: 20%;
-}
-
-.messagefield{
-    display: inline-block;
-    width: 100%;
-    height: 70vh;
-    flex-grow: 1;
-    background-color: white;
-    overflow-y: scroll;
-}
-
-input[type=submit] {
-    visibility: collapse;
-}
-
-input {
-    font: 20px/24px sans-serif;
-    border: 0;
-    outline: none;
-    margin-bottom: 0;
-    
-    width: 95%;
-}
-
-.bottom{
-    position: sticky;
-    bottom: 0;
-    display:flex;
-    width: 100%;
-    left: 0;
-    height: 19vh;
-}
-.material-icons1{
-    background-color: inherit;
-    color: #ffffff;
-    border: 0;
-    padding: 10px;
-    cursor: pointer;
-    position: fixed; /* Фиксированное положение */
-    right: 10px; /* Расстояние от правого края окна браузера */
-}
-
-.attachement{
-    background-color: white;
-    color: #ffffff;
-    border: 0;
-    padding: 10px;
-    cursor: pointer;
-}
-.material-icons1:hover {
-    opacity: 0.8;
-}
-.chat-form {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    border-top: 1px solid #bbbbbb;
-}
