+++ simple-chat/src/yen.html
@@ -178,7 +178,7 @@
                 <button class="material-icons" onclick="window.location.href='/chat_list.html';">
                     <i class="material-icons">arrow_back</i>
                 </button>
-                <div class="img-head"><img src="/src/photos/Yennifer.jpg"></div>
+                <div class="img-head"><img src="https://i.pinimg.com/736x/a9/d3/d0/a9d3d05e702fd49ab554e034288e8fd6.jpg"></div>
                 
                 <div class="status">
                     <h2>Йеннифер</h2>
@@ -198,7 +198,7 @@
 
 
                 <div class="message message-left">
-                    <div class = "btalk-im"> <img src = "/src/photos/Yennifer.jpg"  /> </div >
+                    <div class = "btalk-im"> <img src = "https://i.pinimg.com/736x/a9/d3/d0/a9d3d05e702fd49ab554e034288e8fd6.jpg"  /> </div >
                     <span class="message-text">Привет, тебе нравится моя новая аватарка?</span>
                     <div class="message-meta">
                         <span class="message-time">8:34</span>
@@ -292,7 +292,7 @@ function createDiv(message){
         new_message_img_div.className = "atalk-im";
         
         let new_message_img = document.createElement('img')
-        new_message_img.src = "/src/photos/Tom.jpg";
+        new_message_img.src = "https://i.pinimg.com/originals/7a/26/8f/7a268f7ce7e63295d5c26922335adda9.jpg";
         
         let new_message = document.createElement('div');
         new_message.className = "message message-right";
