+++ simple-chat/webpack.config.js
@@ -12,14 +12,11 @@ const BUILD_PATH = path.resolve(__dirname, 'build');
 module.exports = {
     context: SRC_PATH,
     entry: {
-        index: './scripts/index.js',
-        chatlist: './scripts/chat_list.js'
+        index: './index.js',
     },
     output: {
         path: BUILD_PATH,
-        publicPath: '',
-        filename: 'scripts/[name].bundle.js',
-        chunkFilename: '[name].bundle.js'
+        filename: 'bundle.js'
     },
     module: {
         strictExportPresence: true,
@@ -46,7 +43,7 @@ module.exports = {
                 ],
             },
             {
-                test: /\.css$/,
+                test: /index\.css$/,
                 include: SRC_PATH,
                 use: [
                     {
@@ -61,22 +58,26 @@ module.exports = {
     },
     plugins: [
         new MiniCSSExtractPlugin({
-            filename: 'styles/[name].css',
-            chunks: ['chat_list']
-        }),
-        new MiniCSSExtractPlugin({
-            filename: 'styles/[name].css',
-            chunks: ['index']
-        }),
-        new HTMLWebpackPlugin({
-            filename: 'index.html',
-            template: './index.html',
-            chunks: ['index']
+            filename: 'style.css',
         }),
         new HTMLWebpackPlugin({
             filename: 'chat_list.html',
-            template: './chat_list.html',
-            chunks: ['chat_list']
+            template: './chat_list.html'
+        }),
+        new HTMLWebpackPlugin({
+            filename: 'index.html',
+            template: './index.html'
+        }),
+        new HTMLWebpackPlugin({
+            filename: 'mike.html',
+            template: './mike.html'
+        }),
+        new HTMLWebpackPlugin({
+            filename: 'yen.html',
+            template: './yen.html'
         })
+
+
+
     ]
 };
